% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_coin_price.R
\name{add_coin_price}
\alias{add_coin_price}
\title{Load historic coin data}
\usage{
add_coin_price(input, coins = "", base_currency = "eur")
}
\value{
Returns a tibble with the following columns:
\itemize{
\item \code{symbol} (character): currency symbol (e.g. USD, EUR...)
\item \code{name} (character): currency name
\item \code{currency_id} (character): currency id to sync with kraken assets (e.g. ZEUR)
}
}
\description{
Load all supported currencies. It uses the \href{https://support.kraken.com/hc/en-us/articles/201893658-Currency-pairs-available-for-trading-on-Kraken}{kraken} cash-to-crypto pairs.Zeitraum von coin zu coin unterschiedlich. fehler erscheinen beim import in pp.
}
\examples{
r <- coin_history()
r
}
