% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_coins.R
\name{all_coins}
\alias{all_coins}
\title{Load all supported Coins}
\usage{
all_coins()
}
\value{
A dataframe with the following columns:
* `timestamp` (POSIXct);
* `coin_id` (character): same as the argument `coin_id`;
* `vs_currency` (character): same as the argument `vs_currency`;
* `price` (double): coin price, as of `timestamp`;
* `total_volume` (double): a 24 hours rolling-window trading volume, as
of `timestamp`;
* `market_cap` (double): market capitalisation, as of `timestamp`.
}
\description{
Load all supported crypto coins from the [geckor](https://github.com/next-game-solutions/geckor) package.
}
\examples{
\donttest{
r <- all_coins()
print(r)
}
}
