% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_currencies.R
\name{all_currencies}
\alias{all_currencies}
\title{Load all supported currencies}
\usage{
all_currencies(filter = "")
}
\arguments{
\item{filer}{A \code{string} to filter currencies for a specific provider (currently only \code{"kraken"}).}
}
\value{
Returns a tibble with the following columns:
\itemize{
\item \code{name} (character): currency name
\item \code{sign} (character): currency special character
\item \code{symbol} (character): international currency symbol (e.g. USD, EUR...)
\item \code{kraken_asset} (character): sync helper for kraken ledger transaction
}
}
\description{
Load all supported currencies. It uses the \href{https://sstoeckl.github.io/crypto2/reference/crypto_list.html}{crypto2 fiat_list()} as source
}
\examples{
r <- all_currencies()
r
}
