% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kraken_deposit.R
\name{kraken_deposit}
\alias{kraken_deposit}
\title{fdgfdgf}
\usage{
kraken_deposit(input, base_currency = "eur", lang = "de")
}
\arguments{
\item{input}{A \code{string} with the path to the exported Kraken ledgers data or a \code{data.frame} with the same strucutre as the kraken ledgers export.}

\item{base_currency}{A \code{string} with the prefered currency in which the coin value will be returned. One of \code{c("eur", "usd")}. If the result will be imported in Portfolio Performance, this should be the same as the base currency set there.}
}
\value{
A tibble with the following columns:
\itemize{
\item \code{date} (date): date of the staking transaction;
\item \code{time} (character) time of the staking transaction;
\item \code{type} (character): staking;
\item \code{asset} (character): Kraken coin abbreviation;
\item \code{amount} (double): amount of staked coins;
\item \code{fee} (double): transaction fee;
\item \code{price} (double): coin price in \code{base_currency}, as of \code{date};
\item \code{currency} (character): currency of coin value;
}
}
\description{
Adds coin values from CoinGecko API to Kraken staking transactions
}
\examples{
ledgers <- tibble::tibble(
  refid = 1:5,
  time = c("2022-04-05 03:24:26", "2022-04-11 17:51:44", "2022-04-12 02:43:49", "2022-04-18 09:07:51", "2022-04-19 02:46:48"),
  type = "deposit",
  asset = "ZEUR"
)

r <- kraken_deposit(ledgers, input_type = "data")
head(r, 5)
}
