% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kraken_ledgers_to_pp.R
\name{kraken_ledgers_to_pp}
\alias{kraken_ledgers_to_pp}
\title{Load all supported currencies}
\usage{
kraken_ledgers_to_pp(
  input,
  base_currency = "eur",
  lang = "de",
  filter = FALSE,
  transfer = FALSE
)
}
\value{
A character string with all currency IDs
}
\description{
Load all supported currencies. It uses the \href{https://support.kraken.com/hc/en-us/articles/201893658-Currency-pairs-available-for-trading-on-Kraken}{kraken} cash-to-crypto pairs.
}
\examples{
r <- all_currencies()
r
}
