% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kraken_staking.R
\name{kraken_staking}
\alias{kraken_staking}
\title{Create import file for staking from Kraken}
\usage{
kraken_staking(file, currency = "EUR")
}
\arguments{
\item{file}{(string): path to the exported ledgers.csv file from Kraken}
}
\value{
A dataframe with the following columns:
* `timestamp` (POSIXct);
* `coin_id` (character): same as the argument `coin_id`;
* `vs_currency` (character): same as the argument `vs_currency`;
* `price` (double): coin price, as of `timestamp`;
* `total_volume` (double): a 24 hours rolling-window trading volume, as
of `timestamp`;
* `market_cap` (double): market capitalisation, as of `timestamp`.
}
\description{
Create import file for staking from Kraken
}
