% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp_create_import.R
\name{pp_create_import}
\alias{pp_create_import}
\title{Create import data for Portfolio Performance}
\usage{
pp_create_import(df, type, file, lang = "de")
}
\arguments{
\item{file}{}
}
\description{
Create import data for Portfolio Performance
}
