% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp_rename_columns.R
\name{pp_rename_columns}
\alias{pp_rename_columns}
\title{Rename columns for Portfolio Performance import}
\usage{
pp_rename_columns(df, lang = "de")
}
\arguments{
\item{lang}{(string):}
}
\description{
Rename columns for Portfolio Performance import
}
