% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_col_value.R
\name{split_col_value}
\alias{split_col_value}
\title{Splits a column value in to multiple columns}
\usage{
split_col_value(input, target, sep = " ", colnames)
}
\arguments{
\item{input}{A \code{data.frame}}

\item{target}{A colname `` where the value should be split.}

\item{sep}{A \code{string} on which the value should be split (default = \code{" "}).}

\item{colnames}{A \code{list} with the new colnames. The \code{target} colname should not be included in this parameter.}
}
\value{
Returns the input data with new columns named by the \code{colnames} parameter
}
\description{
Splits a column value on a passed character and stores each element in separate columns.
}
\examples{
data <- data.frame(
  timestamp = c("2022-04-05 03:24:26", "2022-04-11 17:51:44", "2022-04-12 02:43:49",
         "2022-04-18 09:07:51", "2022-04-19 02:46:48")
)

r <- split_col_value(input = data, target = timestamp, colnames = c("date", "time"))
r
}
