% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_csv.R
\name{write_csv}
\alias{write_csv}
\title{Write to local csv file}
\usage{
write_csv(df, output)
}
\arguments{
\item{df}{\code{data} which is stored in csv file}

\item{output}{A \code{string} with the path and filename}
}
\description{
Write data to a local csv file
}
