% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_coins.R
\name{all_coins}
\alias{all_coins}
\title{Load all supported Coins}
\usage{
all_coins()
}
\value{
A tibble with three columns:
\itemize{
\item \code{coin_id} (character): coin IDs, ordered alphabetically;
\item \code{symbol} (character): coin symbols;
\item \code{name} (character): common names of the coins;
}
}
\description{
Load all supported crypto coins from the CoinGecko API via the \href{https://github.com/next-game-solutions/geckor}{geckor} package.
}
\examples{
r <- all_coins()
head(r, 10)
}
