% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_coins.R
\name{all_coins}
\alias{all_coins}
\title{Load all supported coins}
\usage{
all_coins()
}
\value{
A tibble with three columns:
\itemize{
\item \code{coin_id} (character): coin IDs, ordered alphabetically;
\item \code{symbol} (character): coin symbols;
\item \code{name} (character): common names of the coins;
}
}
\description{
Load all supported crypto coins. It uses the \href{https://support.kraken.com/hc/en-us/articles/201893658-Currency-pairs-available-for-trading-on-Kraken}{kraken} cash-to-crypto pairs.
}
\examples{
r <- all_coins()
head(r, 10)
}
