% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_currencies.R
\name{all_currencies}
\alias{all_currencies}
\title{Load all supported currencies}
\usage{
all_currencies()
}
\value{
A character string with all currency IDs
}
\description{
Load all supported currencies. It uses the \href{https://support.kraken.com/hc/en-us/articles/201893658-Currency-pairs-available-for-trading-on-Kraken}{kraken} cash-to-crypto pairs.
}
\examples{
r <- all_currencies()
r
}
