% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp_rename_columns.R
\name{pp_rename_columns}
\alias{pp_rename_columns}
\title{Rename columns for Portfolio Performance import}
\usage{
pp_rename_columns(df, lang = "de")
}
\arguments{
\item{df}{A \code{tibble.} Currently optimized for Kraken ledgers export.}

\item{lang}{Currently only \code{"de"}.}
}
\description{
Translates tibble column names for Portfolio Performance import. Columns that will be renamed: \code{date}, \code{time}, \code{amount}, \code{price}, \code{symbol}, \code{type}, \code{fee}
}
